Split the webpack config file in prod and dev
