+++ package.json
@@ -21,10 +21,11 @@
   "license": "MIT",
   "main": "webpack.config.js",
   "scripts": {
-    "build": "webpack --config webpack.config.js --progress",
+    "build:prod": "webpack --env=prod -p --config webpack.config.js --progress",
+    "build:dev": "webpack --env=dev --config webpack.config.js --progress",
     "lint": "node ./node_modules/eslint/bin/eslint 'src/**/*.js'",
     "lint-fix": "node ./node_modules/eslint/bin/eslint 'src/**/*.js' --fix",
-    "serve": "webpack-dev-server --config webpack.config.js --open"
+    "serve": "webpack-dev-server --env=dev --config webpack.config.js --open"
   },
   "dependencies": {},
   "devDependencies": {
@@ -36,17 +37,17 @@
     "babel-preset-es2017": "^6.24.1",
     "clean-webpack-plugin": "^0.1.16",
     "css-loader": "^0.28.4",
-    "eslint": "^4.2.0",
+    "eslint": "^4.4.1",
     "extract-text-webpack-plugin": "^3.0.0",
     "favicons-webpack-plugin": "0.0.7",
     "file-loader": "^0.11.2",
-    "html-webpack-plugin": "^2.29.0",
+    "html-webpack-plugin": "^2.30.1",
     "html-webpack-template": "^5.6.0",
     "less": "^2.7.2",
     "less-loader": "^4.0.5",
     "style-loader": "^0.18.2",
-    "uglify-js": "^3.0.24",
-    "webpack": "^3.1.0",
-    "webpack-dev-server": "^2.5.1"
+    "uglify-js": "^3.0.27",
+    "webpack": "^3.5.4",
+    "webpack-dev-server": "^2.7.1"
   }
 }
