+++ webpack.config.js
@@ -1,130 +1,3 @@
-const babelPresets = ['es2015', 'es2016', 'es2017'];
-const path = require('path');
-const webpack = require('webpack');
-const CleanWebpackPlugin = require('clean-webpack-plugin');
-const ExtractTextPlugin = require('extract-text-webpack-plugin');
-const FaviconsWebpackPlugin = require('favicons-webpack-plugin');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-
-module.exports = {
-  devServer: {
-    contentBase: path.join(__dirname, 'dist'),
-    compress: true,
-    port: 9000
-  },
-  entry: {
-    app: './src/app.js',
-    print: './src/print.js'
-  },
-  output: {
-    filename: '[name].[chunkhash].min.js',
-    path: path.resolve(__dirname, 'dist')
-  },
-  module: {
-    rules: [
-      {
-        test: /\.js$/,
-        include: /src/,
-        exclude: /node_modules/,
-        use: {
-          loader: 'babel-loader',
-          options: {
-            presets: babelPresets,
-            cacheDirectory: 'cache'
-          }
-        }
-      },
-      {
-        test: /\.less$/,
-        use: ExtractTextPlugin.extract({
-          fallback: 'style-loader',
-          use: [{
-            loader: 'css-loader'
-          }, {
-            loader: 'less-loader'
-          }]
-        })
-      },
-      {
-        test: /\.css$/,
-        use: ExtractTextPlugin.extract({
-          fallback: 'style-loader',
-          use: ['css-loader']
-        })
-      },
-      {
-        test: /\.(gif|jpg|png|svg)$/,
-        use: [
-          'file-loader'
-        ]
-      },
-      {
-        test: /\.(eot|otf|ttf|woff|woff2)$/,
-        use: [
-          'file-loader'
-        ]
-      }
-    ]
-  },
-  plugins: [
-    new webpack.optimize.CommonsChunkPlugin({
-      name: 'common'
-    }),
-    new webpack.DefinePlugin({
-      'process.env': {
-        'NODE_ENV': JSON.stringify('production')
-      }
-    }),
-    new webpack.LoaderOptionsPlugin({
-      minimize: true,
-      debug: false
-    }),
-    new webpack.optimize.UglifyJsPlugin({
-      beautify: false,
-      mangle: {
-        screw_ie8: true,
-        keep_fnames: true
-      },
-      compress: {
-        screw_ie8: true
-      },
-      comments: false
-    }),
-    new CleanWebpackPlugin(['dist']),
-    new ExtractTextPlugin('[chunkhash].min.css'),
-    new FaviconsWebpackPlugin({
-      logo: './assets/images/pingoo.png',
-      persistentCache: true,
-      background: '#3737c8',
-      icons: {
-        android: true,
-        appleIcon: true,
-        appleStartup: true,
-        coast: false,
-        favicons: true,
-        firefox: true,
-        opengraph: false,
-        twitter: false,
-        yandex: false,
-        windows: true
-      }
-    }),
-    new HtmlWebpackPlugin({
-      inject: false,
-      template: require('html-webpack-template'),
-      lang: 'en',
-      meta: [
-        {
-          name: 'description',
-          content: 'A better default template for html-webpack-plugin.'
-        }
-      ],
-      minify: {
-        removeComments: true,
-        collapseWhitespace: true
-      },
-      mobile: true,
-      title: 'My ES6 application skeleton'
-    })
-  ]
+module.exports = function(env) {
+  return require(`./webpack.${env}.js`)
 };
